+++ packages/toprf-secure-backup/src/interfaces.ts
@@ -46,17 +46,14 @@ export type NodeDetailsOverride = {
   pubKeys?: INodePub[];
 };
 
-export type GroupedAuthConnectionParams = {
-  idTokens: string[];
-  authConnectionId: string;
-};
-
 /**
  * AuthenticateParams - The parameters for the authenticate request.
  *
  * idTokens - The idTokens to be used for the authenticate request.
  *
  * authConnectionId - The auth connection name to be used for the authenticate request.
+ *
+ * groupedAuthConnectionId - Optional grouped auth connection id to be used for the authenticate request with aggregate (single id) verifier.
  *
  * userId - The user id of the user to be used for the authenticate request.
  *
@@ -69,7 +66,7 @@ export type AuthenticateParams = {
   idTokens: string[];
   authConnectionId: string;
   userId: string;
-  groupedAuthConnectionParams?: GroupedAuthConnectionParams;
+  groupedAuthConnectionId?: string;
 };
 
 /**